+++ Notificare/NotificarePushLib/NotificareEngine.h
@@ -96,5 +96,9 @@ typedef void (^FileUploadResponseBlock)(NSDictionary*file);
 -(NotificareMKNetworkOperation*)getProductByIdentifier:(NSString *)identifier completionHandler:(NotificationResponseBlock)info errorHandler:(NotificareMKNKErrorBlock) errorBlock;
 
 -(NotificareMKNetworkOperation*)doPurchase:(NSString *)receipt withPrice:(NSString *)price withCurrency:(NSString *)currency forDevice:(NSString *)device completionHandler:(NotificationResponseBlock)info errorHandler:(NotificareMKNKErrorBlock) errorBlock;
+-(NotificareMKNetworkOperation*)getDeviceInbox:(NSString *)deviceId since:(NSString *)date skip:(NSString *)skip limit:(NSString *)limit completionHandler:(NotificationResponseBlock)info errorHandler:(NotificareMKNKErrorBlock) errorBlock;
+-(NotificareMKNetworkOperation*)getInboxMessage:(NSString *)inboxId completionHandler:(NotificationResponseBlock)info errorHandler:(NotificareMKNKErrorBlock) errorBlock;
+-(NotificareMKNetworkOperation*)deleteInboxMessage:(NSString *)inboxId completionHandler:(NotificationResponseBlock)info errorHandler:(NotificareMKNKErrorBlock) errorBlock;
+-(NotificareMKNetworkOperation*)clearInbox:(NSString *)deviceId completionHandler:(NotificationResponseBlock)info errorHandler:(NotificareMKNKErrorBlock) errorBlock;
 
 @end